+++ packages/scan/src/web/utils/log.ts
@@ -3,7 +3,7 @@ import { ChangeReason, type Render } from "~core/instrumentation";
 import { getLabelText } from "~core/utils";
 import { REACT_SCAN_PRO_LOG_PREFIX } from "../../logging-constants";
 
-export const log = (renders: Array<Render>) => {
+export const log = (renders: Array<Render>, namespace: "notification" | "session-report") => {
   const logMap = new Map<
     string,
     Array<{ prev: unknown; next: unknown; type: string; unstable?: boolean }>
@@ -69,12 +69,19 @@ export const log = (renders: Array<Render>) => {
   for (const [name, changeLog] of Array.from(logMap.entries())) {
     // oxlint-disable-next-line no-console
     console.group(
-      `${REACT_SCAN_PRO_LOG_PREFIX} %c${name}`,
+      `${REACT_SCAN_PRO_LOG_PREFIX} [${namespace}] %c${name}`,
       "background: hsla(0,0%,70%,.3); border-radius:3px; padding: 0 2px;",
     );
     for (const { type, prev, next, unstable } of changeLog) {
       // oxlint-disable-next-line no-console
-      console.log(REACT_SCAN_PRO_LOG_PREFIX, `${type}:`, unstable ? "⚠️" : "", prev, "!==", next);
+      console.log(
+        `${REACT_SCAN_PRO_LOG_PREFIX} [${namespace}]`,
+        `${type}:`,
+        unstable ? "⚠️" : "",
+        prev,
+        "!==",
+        next,
+      );
     }
     // oxlint-disable-next-line no-console
     console.groupEnd();